pyproject.toml
[tool.poetry]
name = "backend-react-tailwind"
version = "0.1.0"
description = ""
authors = ["John Ngugi <johnngugi0407@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.35"
geopandas = "^1.0.1"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
